figure out auth (#25)
Auth will work via ownership, gated by permissions to create an
ownership. Once ownership is granted, only admins and the owner can
change permissions.

Permissions are:

- signal: allowed to send signals to the resource
- completion: allowed to be notified of completion
- output: allowed to get output
- call: allowed to call
- lock: allowed to lock
- share+: allowed to create new shares
- share-: allowed to remove shares and view other shares

These permissions can be shared with individual users or roles if the
resource authz mode is set to "explicit."

If the mode is "auth," then any authenticated user is granted full
permissions to the resource. For more fine-grained permissions, the
developer should set a role and grant that role to the specific
permissions they desire.

If the mode is "anon," then any user on the internet is granted full
permissions to the resource.

The default is "explicit" with no other permissions (only the owner can
manage the permissions).